# Optimize model matrix calculations