… docs/ScriptingReference/ClassReference.md
… src/Engine/Manager/RenderManager.cpp
… src/Engine/Manager/ScriptManager.cpp
… src/Engine/Manager/SoundManager.cpp
… src/Engine/Util/MousePicking.cpp