Write a specialized function to get the model matrix directly instead of calculating matrices for scaling, rotation and translation independently and then multiplying them together. Also cache matrices and only update them if they are dirty.

In a Sponza test scene, measured on an AMD Ryzen 5 2600X, setting up the render scene goes from ~0.32 ms to ~0.10 ms.
On a Samsung Galaxy S21 (Exynos), the same scene goes from ~12.06 ms to ~0.12 ms.